requirements.txt
aiofiles==22.1.0; python_version >= "3.7" and python_version < "4.0"
anyio==3.6.1; python_full_version >= "3.6.2" and python_version >= "3.7"
asyncpg==0.26.0; python_full_version >= "3.6.0"
click==8.1.3; python_version >= "3.7"
colorama==0.4.5; python_version >= "3.7" and python_full_version < "3.0.0" and platform_system == "Windows" or platform_system == "Windows" and python_version >= "3.7" and python_full_version >= "3.5.0"
fastapi==0.84.0; python_version >= "3.7"
h11==0.13.0; python_version >= "3.7"
idna==3.4; python_full_version >= "3.6.2" and python_version >= "3.7"
jinja2==3.1.2; python_version >= "3.7"
markupsafe==2.1.1; python_version >= "3.7"
numpy==1.23.3; python_version >= "3.10"
pandas==1.4.4; python_version >= "3.8"
phonenumbers==8.12.55
pydantic==1.10.2; python_version >= "3.7"
python-dateutil==2.8.2; python_version >= "3.8" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.8"
python-dotenv==0.21.0; python_version >= "3.7"
pytz==2022.2.1; python_version >= "3.8"
six==1.16.0; python_version >= "3.8" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.8"
sniffio==1.3.0; python_full_version >= "3.6.2" and python_version >= "3.7"
starlette==0.19.1; python_version >= "3.7"
types-aiofiles==22.1.0
types-ujson==5.4.0
typing-extensions==4.3.0; python_version >= "3.7"
uvicorn==0.18.3; python_version >= "3.7"
